# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Test task",
        "title": "Money Processing Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/processing",
    "paths": {
        "/account": {
            "post": {
                "description": "Create account with the specified currency for the client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creating new account",
                "parameters": [
                    {
                        "description": "Account Data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.newAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New account UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    }
                }
            }
        },
        "/account/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves account based on given ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.outAccount"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    }
                }
            }
        },
        "/client": {
            "post": {
                "description": "Create new client with specified name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creating new client",
                "parameters": [
                    {
                        "description": "Name for the client",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.newClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New client UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    }
                }
            }
        },
        "/client/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves client based on given ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.outClient"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    }
                }
            }
        },
        "/transaction": {
            "post": {
                "description": "Create transaction based on transaction type, account IDs and transaction amount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creating transaction",
                "parameters": [
                    {
                        "description": "Transaction data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.newTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New transaction UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    }
                }
            }
        },
        "/transaction/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves transactions list based on given account ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/http.outTransaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.errResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.errResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "http.newAccount": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "currency_id": {
                    "type": "integer"
                }
            }
        },
        "http.newClient": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "http.newTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "source_id": {
                    "type": "string"
                },
                "target_id": {
                    "type": "string"
                },
                "type_id": {
                    "type": "integer"
                }
            }
        },
        "http.outAccount": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "currency": {
                    "$ref": "#/definitions/http.outCurrency"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "http.outClient": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/http.outAccount"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "http.outCurrency": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "http.outTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                },
                "target_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}